Now the actual bug fix.
Problem was that in balance-sheet & balance-by-date reports,
if we had an account that had a bunch of single children, the last of
which had >1 children, we would look for those >1 children at the top
account. Now we use the (collapsed) name in the tree to dig down to
the last node with the collapsed name instead of the first node in the
collapsed name before we look for the children.

(Not very clear, is it?)

Bug was actually fixed in balances-by-date report just through adding the
extra type information. Have now added the fix into balance-sheet report
also.